Extract `Cue` component into file
