+++ packages/story-editor/src/components/canvas/videoCaptions/trackRenderer.js
@@ -30,7 +30,7 @@ import styled from 'styled-components';
  * Internal dependencies
  */
 import { useLayout } from '../../../app';
-import parseTimestamp from './parseTimestamp';
+import Cue from './cue';
 
 const Track = styled.div`
   width: 100%;
@@ -40,87 +40,7 @@ const Track = styled.div`
   font-size: ${({ $fontSize }) => `${$fontSize}px`};
 `;
 
-const CueWrapper = styled.span`
-  text-align: center;
-  height: ${({ $height }) => `${$height}px`};
-`;
-
-const CueEl = styled.span`
-  color: ${({ $isFutureCue, theme }) =>
-    $isFutureCue ? theme.colors.fg.black : theme.colors.fg.white};
-`;
-
-// TODO: Adjust font size and styling based on canvas size.
-const Section = styled.span`
-  margin: 0 10px 10px;
-  padding: 6px 12px;
-  vertical-align: middle;
-  border-radius: 15px;
-  background: rgba(11, 11, 11, 0.6);
-  color: rgba(255, 255, 255, 1);
-  display: inline-block;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-    Oxygen-Sans, Ubuntu, Cantarell, 'Helvetica Neue', sans-serif;
-  line-height: 1.4;
-  word-break: break-word;
-  word-wrap: break-word;
-  overflow-wrap: break-word;
-`;
-
 /** @typedef {import('react').ReactNode} ReactNode */
-
-function Cue({ cue, videoTime, height }) {
-  const html = cue.getCueAsHTML();
-
-  return (
-    <CueWrapper $height={height}>
-      <Section>
-        {[...html.childNodes].map((node) => {
-          if (node.target === 'timestamp') {
-            return null;
-          } else {
-            let timestamp = null;
-            if (node.previousSibling?.target === 'timestamp') {
-              timestamp = parseTimestamp(node.previousSibling.data);
-            }
-            const isFutureCue = timestamp > videoTime;
-
-            const isText = node.nodeType === Node.TEXT_NODE;
-
-            if (isText) {
-              return timestamp !== null ? (
-                <span>
-                  <CueEl $isFutureCue={isFutureCue}>{node.textContent}</CueEl>
-                </span>
-              ) : (
-                <CueEl>{node.textContent}</CueEl>
-              );
-            }
-
-            return timestamp !== null ? (
-              <span>
-                <CueEl
-                  $isFutureCue={isFutureCue}
-                  dangerouslySetInnerHTML={{ __html: node.innerHTML }}
-                />
-              </span>
-            ) : (
-              <CueEl dangerouslySetInnerHTML={{ __html: node.innerHTML }} />
-            );
-          }
-        })}
-      </Section>
-    </CueWrapper>
-  );
-}
-
-Cue.propTypes = {
-  cue: PropTypes.shape({
-    getCueAsHTML: PropTypes.func,
-  }).isRequired,
-  videoTime: PropTypes.number.isRequired,
-  height: PropTypes.number.isRequired,
-};
 
 /**
  *
